feat: relocate test to proper location
